Add boilerplate and structure the content
The banner consists of a header and hero. The header includes the logo and links
while the hero has a sign up card with a picture to its right.

The upper contents will display four apps and a description beneath each.

The lower contents section will include a quote.

The sub-section will have a sign up card

The footer will have the logo and date of inception