Add and format banner image and button
The button was give a transition, hover effect, and shadow to help the
user identify when the cursor is over the button vs clicking the
button.

The image was given a min-width of 400px to be consistent with the sign
up card in the banner. An aspect ratio was used to adjust the size of
the picture with the size of the viewport. The image was cropped using
object-fit and object-position to focus on the object of the image.