Format the content section with a title and apps
This section will be used to display the available apps for download by
the developer as well as give a sneak peak to the unpublished apps.
Underneath each image of an app will be a short description.

This section uses flexbox and uses space-between to help make the layout
easily readible to the user.