Add padding to content
Finished styling the content section of the page by adding padding. The
padding helps to make the information about the apps front and center,
making the info and visuals easier to focus on. The separation between
the apps portion and the quote, helps the user to digest the
the information in chunks rather then being overwhelmed or confused by
the clutter.