Style the quote in the sub-content section
The sub-section has padding to make it easier to read. It was made
italic to clearly indicate the text is a quote. The source fo the quote
was put beneath the text towards the right in order to credit the
the source and give the reader another cue that the text is a quote.